+++ layabase/testing.py
@@ -1,3 +1,8 @@
+import datetime
+
+import pytest
+
+
 def reset(base):
     """
     If the database was already created, then drop all tables and recreate them all.
@@ -12,3 +17,43 @@ def reset(base):
         import layabase.database_sqlalchemy as database_sqlalchemy
 
         database_sqlalchemy._reset(base)
+
+
+_date_time_for_tests = datetime.datetime(2018, 10, 11, 15, 5, 5, 663979)
+
+
+class DateTimeModuleMock:
+    class DateTimeMock:
+        @staticmethod
+        def utcnow():
+            return _date_time_for_tests
+
+    datetime = DateTimeMock
+
+
+@pytest.fixture
+def mock_mongo_health_datetime(monkeypatch):
+    import layabase.database_mongo
+
+    monkeypatch.setattr(layabase.database_mongo, "datetime", DateTimeModuleMock)
+
+
+@pytest.fixture
+def mock_mongo_audit_datetime(monkeypatch):
+    import layabase.audit_mongo
+
+    monkeypatch.setattr(layabase.audit_mongo, "datetime", DateTimeModuleMock)
+
+
+@pytest.fixture
+def mock_sqlalchemy_health_datetime(monkeypatch):
+    import layabase.database_sqlalchemy
+
+    monkeypatch.setattr(layabase.database_sqlalchemy, "datetime", DateTimeModuleMock)
+
+
+@pytest.fixture
+def mock_sqlalchemy_audit_datetime(monkeypatch):
+    import layabase.audit_sqlalchemy
+
+    monkeypatch.setattr(layabase.audit_sqlalchemy, "datetime", DateTimeModuleMock)